Simplified the Dockerfile.x86.  The other files changed as a consequence of this simplification.
I realized that I had started to put a proxy server to do the 2 way tls
authentication within the container and had decided it was simpler to have
the proxy server in a different container.

I also started my docker container using centos as it is similar to RHEL.
However this one starts with the python 3.8 image, which runs under a different
linux distribution.