Remove unnecessary check for content-length header
I noticed that requests to PUT /projects/<project_name> were failing
when using a tool other than curl, and that's because it was setting
"Content-length: 0" rather than omitting the content-length header,
causing the service to crash.

This removes the unnecessary check for the content-length header;
Flask takes care of that for us.