Use an application factory to create Flask app
By moving application creation into the create_app() method, we are
able to isolate application instantiation from the import of the wsgi
module. This is going to be necessary for writing unit tests, because
we will often want to mock out some top-level objects that can only be
modified after the module has been imported.

The pattern used in this PR is called an "application factory", and is
explained in more detail in [1].

[1]: https://flask.palletsprojects.com/en/2.0.x/patterns/appfactories/