Allow authentication to be disabled at run time
This adds an AUTH_DISABLED configuration item to the flask app which
allows us to disable authentication. We can use this during
development to make it easier to test the service, and we can use it
during testing to create a test client that doesn't require
authentication with every call.