Create a kubernetes client using requests.Session
Moves all the logic for communicating with the Kubernetes API into the
kubeclient module. This replaces the get_request, del_request,
put_request, and post_request methods of MocOpenShift with calls to
the session object, which takes care of things like setting the
appropriate headers.

This also enables SSL certificate validation by default when run in
Kubernetes.