Simplify Dockerfile
As we add additional files to the project it becomes inconvenient to
constantly modify the Dockerfile to include them. This updates the
Dockerfile to copy everything in the current directory, and adds a
.dockerignore file to prevent things that change regularly (like the
.git directory) from invalidating the cache.

NB: when building locally this will pick up anything in the current
working directory. This isn't an issue when building in CI or via the
automatic build features of quay.io or docker hub.