Use application factory pattern to create flask app (#62)
* Use an application factory to create Flask app

By moving application creation into the create_app() method, we are
able to isolate application instantiation from the import of the wsgi
module. This is going to be necessary for writing unit tests, because
we will often want to mock out some top-level objects that can only be
modified after the module has been imported.

The pattern used in this PR is called an "application factory", and is
explained in more detail in [1].

[1]: https://flask.palletsprojects.com/en/2.0.x/patterns/appfactories/

* Replace MocOpenShiftSingleton with single instance

Instead of creating a new MocOpenShfit4x instance in each handler
function, just create one when we start up the app. This allows us to
drop the MocOpenShiftSingleton wrapper.

x-branch: feature/application-factory