Prepare for tests (#63)
* Create a kubernetes client using requests.Session

Moves all the logic for communicating with the Kubernetes API into the
kubeclient module. This replaces the get_request, del_request,
put_request, and post_request methods of MocOpenShift with calls to
the session object, which takes care of things like setting the
appropriate headers.

This also enables SSL certificate validation by default when run in
Kubernetes.

* Simplify Dockerfile

As we add additional files to the project it becomes inconvenient to
constantly modify the Dockerfile to include them. This updates the
Dockerfile to copy everything in the current directory, and adds a
.dockerignore file to prevent things that change regularly (like the
.git directory) from invalidating the cache.

NB: when building locally this will pick up anything in the current
working directory. This isn't an issue when building in CI or via the
automatic build features of quay.io or docker hub.

* Use flask config to hold config information

Use the config attribute of the flask app to aggregate configuration
from defaults, explicitly provided parameters, and the environment.
This provides a consistent mechanism for accessing values, and makes
it easier to pass in explicit config information when writing tests.

* Allow authentication to be disabled at run time

This adds an AUTH_DISABLED configuration item to the flask app which
allows us to disable authentication. We can use this during
development to make it easier to test the service, and we can use it
during testing to create a test client that doesn't require
authentication with every call.

* Store quota in repo as a plain JSON file

By using a configmapGenerator [1] to generate the quota ConfigMap, we can
store the example data itself as a plain JSON file, which makes it
available to the code when running things locally.

[1]: https://kubernetes.io/docs/tasks/manage-kubernetes-objects/kustomization/#configmapgenerator

* Avoid long sleeps as a synchronization mechanism

Using `sleep` as a synchronization mechanism is problematic -- it
can both introduce unnecessary delays and result in unexpected
problems if an operation takes longer than expected.

This commit replaces the use of `sleep` with explicit polling.

* Create acct_mgt package for Python sources

Move all the account management code into the acct_mgt package.
Putting the code into a single namespace like this makes it easier to
write modular code, keeps the top-level directory organized, and makes
testing simpler.

x-branch: feature/prepare-for-tests