Unit tests and new functional tests (#65)
* Introduce new functional tests

This introduces a new suite of functional tests. The tests are moved
into the `tests/functional` directory, which will allow us in
subsequent commits to place unit tests in `tests/unit`.

* Add unit tests for flask app

This adds a suite of unit tests for the flask application and modifies
the CI workflow to run unit tests in addition to the pre-commit
checks.

* Add unit tests for moc_openshift quota functions

Several tests are marked with @pytest.mark.xfail() to indicate places
where the existing API may be problematic. For each of these tests, we
need to determine if the existing behavior is correct, and if it is,
modify the test so that it tests the current behavior and passes.

* Add unit tests for moc_openshift user functions

* Add unit tests for identity operations

* Add tests for acct_mgt.kubeclient

* Add unit test workflow

Add a GitHub workflow to run unit tests for each push and
pull-request.

* Add support for coverage reporting

- Add pytest-coverage to test-requirements.txt, which allows us to
  produce coverage reports.

- Update the CI workflow to include coverage reporting, which looks
  like this:

      Name                        Stmts   Miss  Cover
      -----------------------------------------------
      acct_mgt/__init__.py            0      0   100%
      acct_mgt/app.py               141     10    93%
      acct_mgt/defaults.py            2      0   100%
      acct_mgt/kubeclient.py         33     19    42%
      acct_mgt/moc_openshift.py     299    133    56%
      acct_mgt/wsgi.py                7      7     0%
      -----------------------------------------------
      TOTAL                         482    169    65%

- Add a note to the README about running unit tests.

x-branch: feature/tests