Set minimum size of a container
This prevents people from explicity setting memory or cpu request to 0 for their
containers. While such pods will be the first to be kicked out when the
cluster is low on resources, they will continue to run for free at other times.

With this in place the scheduler will reject the container from running with
an appropriate message to the user.